+++ touchables/TouchableOpacity.web.js
@@ -1,0 +1,2 @@
+import { TouchableOpacity } from 'react-native';
+export default TouchableOpacity;
